+++ Changelog.md
@@ -1,5 +1,5 @@
 ## 0.0.26
-*INCOMPLETE*
++ Added support for Anthropic `thinking` and `output_config.effort` in `muna.beta.anthropic.messages.create`.
 
 ## 0.0.25
 + Added support for specifying `seed` in `muna.beta.openai.chat.completions.create`.